[JENKINS-38433] Make the agent section pluggable.
This is very preliminary - I expect a lot more iteration. But it
works. The DeclarativeAgent* classes are in a separate plugin so that
they can be depended on by other plugins with minimal transitive
dependencies in the process (just workflow-cps-plugin currently - may
try to find a way to narrow that down further).