Normalizing code block string toGroovy and equals.
Until this commit, we were over-indenting any line after the first in
a multi-line code block (i.e., `script`, `when`). That is, we were
retaining the original indentation for the second, third, etc lines
and then adding the normal indentation we'd add anyway. That...broke
stuff. Most notably, it broke round-trip conversion. So now, we trim
away existing indentation in toGroovy() *and* ignore indentation
differences in .equals().