Get rid of ModelASTWhenCondition marker
The only reason to keep it would be to keep the
validateElement(ModelASTWhatever) method consistency, and frankly
that's not enough. What we need is to be able to call a different
validation method when a ModelASTStep is being validated within a
ModelASTWhen than in other contexts, so, well, let's add
validateWhenCondition(ModelASTStep) instead of relying on class magic.