[FIXED JENKINS-40984] Always evaluate/run all post conditions
Even if there's an error in an earlier condition execution, continue
to the subsequent ones. Additionally, switch to doing a one-off check
at the beginning of the post section to see if *any* conditions are
satisfied at that time, and then when actually iterating through the
conditions, check for satisfaction at *that* time. That's so that, for
example, if the build is successful when it gets to evaluating post
conditions and then there's a failure in the execution of `always`,
the `success` block won't be executed, and the `failure` block will,
since the build status has changed.