[FIXED JENKINS-41748] Handle env var cross references properly
This changes things to now properly expand environment variable names
referred to in `environment` values, building from the job context's
environment to the parent context (if per-stage `environment`) and
lastly to the immediate `environment` context.

This is not exactly *perfect*, per se - you still need to either
single quote or escape the dollar sign due to the `GString` getting
resolved before we can populate anything. May revisit that in the
future to see if there's a workaround we can use.