[FIXED JENKINS-41890] Resolve env after agent
This is to make sure we have `WORKSPACE` in the environment. Note that
this won't quite work right for per-stage agent/environment, though -
`WORKSPACE` there will still refer to the top-level `WORKSPACE` (if
there is one) or be null if `agent none` is used at the top
level. This is so that we have the per-stage environment populated
before evaluating any when condition, while not launching in a
per-stage agent until said when condition has already been
evaluated. Might reconsider this in the future.

Also added `.logMatches(String... patterns)` to
`AbstractModelDefTest.ExpectationsBuilder` so that we can do regular
expression matching against logs, which we needed to verify this.