Backwards compat for running Pipelines
Good point made by Bobby - we'd been bitten in the past when we
changed the API for methods called from ModelInterpreter, since
ModelInterpreter will already be loaded in running Pipelines at
upgrade time, resulting in no such method errors. So here I add a
deprecated backwards-compatible method for the old getDeclarativeAgent
signature, along with null checking for root in getDeclarativeAgent
for that scenario.