Followup to JENKINS-41900 - just record if in stage
There's no actual good reason for carrying the whole context (i.e.,
Root or Stage object) around in the DeclarativeAgent when all we need
to know is whether we're already in a stage. So...switch that to a
boolean and have Agent.groovy check whether the context it's handed is
a Root instance. If it is a Root instance, we're not in a stage. If it
isn't a Root instance, we are in a stage. Tada.