[JENKINS-42640] Fix validation of wrappers and some types
In validating the options contents, we didn't actually look for a
StepDescriptor, just a symbol-driven Descriptor. Which won't work for
steps! So we never actually validated the contents of wrappers in
options. Besides that, we needed to reject converting to int from
String/Boolean - that gets through ScriptByteCodeAdapter.castToType,
but not runtime.