[FIXED JENKINS-42551] Reject mismatched quotes and other invalid Groovy
So now the JSON string "hello\\'" will cause a parse-time error. I
also tweaked JSON validation endpoint to also convert to Jenkinsfile
and validate *that*, so as to be sure we've actually got JSON that can
be a valid Jenkinsfile, even if it has some kind of gotcha in it like
the one fixed here.