[FIXED JENKINS-42771] Allow + binary expressions in env values
This worked previously because we didn't actually pay attention to
what was really in the value for an environment variable in the AST
model, only at runtime. And the runtime parsing worked differently, so
it did fine with `FOO = "A" + "B" + "C"`. Well, now so does the AST
model. Yay!