[FIXED JENKINS-42829] Allow references outside pipeline block to work
Now that we're script.evaluate(...)ing everything, we need to be sure
that anything defined elsewhere in the Jenkinsfile outside the
pipeline block is available in that evaluated scope. This...is the
most effective way I can come up with to do so. It has some risks -
something with side effects could be problematic, but that's not a use
case we're keen on supporting anyway.