[FIXED JENKINS-41456] Properly validate publishHTML
Specifically, when we see that there's a sole required parameter, the
only provided argument by the user is a map, at least one of those map
keys doesn't have a corresponding parameter in the describable, and
the sole required parameter's type is a describable itself, validate
the argument map against that parameter type describable instead of
against the parent describable. So for publishHTML, that means
realizing `target` is an `HtmlPublisherTarget` and validating the map
against that.