[FIXED JENKINS-43035] Add imports option for libraries
Also, just add libraries and imports directly to the AST. This doesn't
solve manual use of the @Library annotation + an import, but the right
answer is to use the libraries directive *anyway*. This also gets rid
of the usage of the library step, amusingly enough.

Note - this is not backwards-compatible in the JSON form, but @kzantow
assures me that the editor is not yet doing anything with libraries,
so we still have that opportunity now anyway, and this format is going
to be preferrable in the long term anyway.