[FIXED JENKINS-42772] Don't fail out for single + required missing param
More specifically, don't fail validation if there's only a single
parameter, it's required, and there are no arguments specified by the
user. This may end up causing some surprise errors at runtime, but it
appears that most everything is assuming they need to do null checks
on just-one-and-it's-required parameters for steps, so let's give this
a shot.