Work around PluginFirstClassLoader+GroovyClassLoader problem
PluginFirstClassLoader doesn't have a functional findResource method,
which for some reason I have yet to determine breaks
GroovyClassLoader.loadClass when trying to fetch a groovy file from a
PluginFirstClassLoader via the UberClassLoader. This works around that
problem by catching the ClassNotFoundException and trying to instead
use getResource to get the URL for the groovy file and then parse it
directly.

Also whitelists the fields on WithScriptScript so they can be accessed
by children outside of these plugins properly.