Add CodeNarc!
Note that we *are* scanning both the compiled and CPS code in
pipeline-model-definition. We're only failing for priority 1 warnings
- priority 2 is a lot of "don't just catch Exception" or "empty
if/else block" that I...frankly don't care about. And most of the
priority 3 are unused imports, which would be nice to clean up, sure
but don't actually *matter* in the grand scheme of things.

Also, opened https://issues.jenkins-ci.org/browse/INFRA-1267 to get
the Warnings plugin (and Violations plugin) installed on ci.jenkins.io
so we can actually report on failures in a useful way.