[FIXED JENKINS-45916] Properly handle overwriting+including existing vars
Technically, JENKINS-45916 is already fixed by earlier changes here,
but in the process we introduced a hellish break due to trying to call
the closure for getting the value of a variable when it's referenced
in *setting* the variable in the first place. So if we're setting FOO
and the value includes FOO, just use the original expression for that
instead of closure-calling there.