Tweak invalidWrapperType tests
Sometimes, when multiple tests are being run and...some other
condition applies (haven't figured out what, since I can't reproduce
it locally), the TestExtension in MultipleUnnamedParametersTest ends
up being present in other tests, most notably ErrorsEndpointOpsTest
and ErrorsJSONParserTest. Since we have to hardcode the expected error
in those tests, without it being calculated at the moment the test's
running, like ValidatorTest#invalidWrapperType, this can result in
failures where the list of valid options the test is expecting to see
is calculated before the TestExtension is added but the actual error
message is generated afterwards and so includes the TestExtension.

Since we have the full check including the list of valid types in
ValidatorTest, we might as well simplify the invalidWrapperType tests
elsewhere by using a shorter and consistent expected error string.