Add test comparing original groovy to original json
I got confused by something and forgot that we didn't have this in the
first place. Had to update a few json files to match due to changes in
the groovy files since the json was created. This should keep us from
hitting a few weird regressions in the future.