[FIXED JENKINS-46359] Reset failure status in retry option
Also change the build condition responder extension point to take an
exception, which we can use to decide what the build status actually
is. Backwards compatibility is retained (see Unstable.groovy and
Always.groovy - they don't need to do anything with the exception, so
I left them as is to demonstrate the backwards compatibility).