[FIXED JENKINS-47197] Make sure we don't end up with Java 8 bits
Groovy does something weird with `@Delegate` on a `Map` with Java 8
but targeting Java 7, resulting in a whole lot of additional methods
created using `java.util.function.Function` and friends. That's
annoying. To say the least.  Took me a while to find where that
reference was being inserted, but yeah, it was `MappedClosure` and its
`@Delegate`, so I got rid of `@Delegate` and now things are fine. Phew.