[FIXED JENKINS-47421] Make sure we init stagesList in readResolve
This is a very minor issue in practice - the ExecutionModelAction
isn't really used by anything but our own check for multiple executed
pipeline blocks, which itself can't ever be the case on builds that
started before Declarative 1.2 in the first place. Buuuuut it does
result in some weird unreadable data messages in Jenkins, so let's
play things very safe.