[FIXED JENKINS-47781] Narrow model lookups when appropriate
The original problem here is that if you just look up describables by
the "upstream" symbol, you'll get one from Copy Artifact before you
get `ReverseBuildTrigger`. Which...ok. Except that our syntax checking
treats kicks in for the one from Copy Artifact! So let's narrow our
model lookups when appropriate.

I don't *love* this solution - it could be smarter and have a better
understanding of context. But this works for now, so I'll go with it.