[FIXED JENKINS-46597] Further handling of serializability issues
Specifically: the checkout step returns a TreeMap, so let's make sure
we never have a variable assigned directly to that return
value. Probably want to change checkout to return a HashMap in the
future, but we'll cross that bridge later.

Also updated workflow-cps dependency so that we can be sure
BooleanClosureWrapper serialization issues won't crop up, letting us
write more things as .every or .any.