[FIXED JENKINS-48209] Prevent serialization of when expression closure
This shouldn't really matter, in truth - I believe we actually
regenerate the whole Root object from scratch on resume, though I
could be misremembering. Regardless, even having the when expression
closure be transient, the durability test (with when expressions
added) still works right.

As to why we need to do this in the first place? Because something
weird could be out there trying to some non-standard serialization of
Describables. Ok, spoiler, there definitely is such a thing, and it
does a straight XStream serialization of every Describable. But we use
Describables as more than just UI config stuff - we use them for
syntax validation and such for our internal extension points, like
when conditions. And if that XStream serialization of a when
expression with a sh step kicks in, well, all hell breaks loose. This
works around that.