[FIXED JENKINS-48266] Skip post for skipped parallel stages
The JENKINS-47928 fix was actually not right - it disabled calling
post entirely for parallel parent stages if they weren't the source of
an error themselves. That means they weren't called at all if there
was no failure!

The proper fix is to track whether a parallel parent stage has been
skipped, and if so, skip post execution as well. Which is what's done
here. Tada.