[JENKINS-49135] Add support for passing Options var/method to options
This is very preliminary (and even the minimal approach still needs
tests for `StageOptions` as well as `Options`) and may not actually
land. The idea is that you'd be able to either declare a variable of
type `Options` or have a method that returns an instance of `Options`,
and use that variable or method call as the argument to `options`,
like this:

```
def opts = new Options([buildDiscarder(logRotator(numToKeepStr:'1'))], [:], [:])

pipeline {
  ...
  options opts
  ...
}

// or

def genOpts() {
    return new Options([buildDiscarder(logRotator(numToKeepStr: '1'))], [:], [:])
}

pipeline {
  ...
  options getOpts()
  ...
}
```

This would *not* allow for validation of the `options` contents, it
*would* require actually instantiating an `Options` instance, rather
than defining a closure like you'd normally specify for `options`, and
it would not work in the editor (i.e., translating to JSON would just
result in an empty `options` directive). This is necessary because the
transformation behind the scenes from `options { ... }` to an
`Options` instance requires runtime AST transformation, which means we
need to treat any `options` method call anywhere in the Jenkinsfile or
a shared library specially. Which is not optimal, to say the least.

If this ever actually lands, it'll be a power-user feature anyway, so
I'd be ok with this requirement. That said, this is a good reason why
I'm unsure about this landing.