[FIXED JENKINS-48752] Drastically simplify and fix changed calculation
The previous was...complicated. In some contexts, we do need to check
the execution result, not just the run result, but that led to me
writing some crappy code that meant that changed always triggered for
any status but SUCCESS. Sigh. So instead, specifically get the worst
status from the run and the execution and use that for
comparison. Suddenly everything's simpler. =)