[JENKINS-42688] More powerful conditions in post
Specfically, allow reusing `when` conditions in `post`. The syntax is
still up for negotiation, but we do need to be able to specify both
the `when` condition and the `steps`, so I think the syntax I've got
is as good as it gets:

```
post {
  condition {
    when {
      allOf {
        branch "master"
        buildStatus status:"SUCCESS"
      }
    }
    steps {
      echo "tada"
    }
  }
}
```

Old-school `success`, `always`, etc will still work as well. To retain
backwards compatibility with some...idiosyncratic behavior, we'll
first check to see if *any* conditions (traditional `post` or the new
`when`-based ones) are satisified, then process the satisfied
`when`-based conditions, and *then* iterate over the traditional
`post` conditions, checking if it's satisified, and then
executing. This is for cases, like, say, doing `junit` in `always` and
wanting to fire something in `unstable`.

Might be worth considering adding an ordinal of some sort to the
`when`-based `conditional`s, so that you can control order of
execution and evaluation there too?

Still needs testing of the new conditions - right now I've just got backwards
compat testing.