[FIXED JENKINS-49558] Add containerPerStage option
If given with a top-level Docker or Dockerfile agent, the top-level
will magically switch to just a label agent, while each stage without
an explicit agent specified will end up using the root agent
definition with reuseNode enabled. The result will be that each stage
gets a new container. Tada.

This could theoretically cause problems in resuming builds after
upgrade, but not likely, since we actually instantiate the
DeclarativeAgent instance just in time, not at the beginning of the
build. So the new field AbstractDockerAgent#containerPerStageRoot
should get populated correctly.