Create our own per-test timeout rule.
JenkinsRule's timeout logic goes batty when JenkinsRule is used as a
ClassRule - it ends up applying to the *entire* test class, not just
the individual test. So that's annoying. Here, I copy-pasted some
stuff around to create a per-test Rule that still does the nice
thread-dump stuff on timeout. This is just a hack, mind you.