[FIXED JENKINS-44993] Don't swallow errors in post
We only throw one exception for a failure in a stage, whether that
failure is in the `steps` block or in a `post` condition's
execution. Any subsequent exceptions get swallowed and never
logged. This will change that, so that we'll start logging any
exceptions encountered in executing `post` blocks even if the stage
has already hit an error before getting to the `post` blocks.