[FIXED JENKINS-50882] Make generator work on multibranch and org folders
Sigh - I didn't realize that Snippetizer actually had to do the same
thing in three places (well, more accurately, two of those are in
workflow-multibranch but are otherwise identical to what's in
Snippetizer). So here are `TransientActionFactory`s added for
`OrganizationFolder` and `WorkflowMultiBranchProject` (aka multibranch
folder), and a refinement of the `WorkflowJob` one too.