[FIXED JENKINS-51872] Fix tool dropdown
SECURITY-624 resulted in the `<f:option>` body going through
`h.xmlUnescape`...which is fine, except that when it's being specified
inside a `l:renderOnDemand` block, the `h` variable isn't captured!
Which results in `h.xmlUnescape('literally anything')` being null. So
let's capture `h` when we call `l:renderOnDemand`, just to be safe.