[FIXED JENKINS-51383] Handle Throwables, not just Exceptions
Previously, we only caught and handled `Exception`s in
`ModelInterpreter`. Which was fine in 99% of cases, but not
`NoSuchMethodError` - that's an `Error`, not an `Exception`, and so it
slipped through the cracks, resulting in `post` still firing for
`success` even though the build would end up failed. So let's fix that.