[FIXED JENKINS-51227] Add quietPeriod DeclarativeOption
This will let you set the quiet period for the job from the options
block. As with job properties, triggers, and parameters, the quiet
period will only be set to the default of 0 if the previous build
actually had the QuietPeriod DeclarativeOption specified and the
current build does not, so that UI-configured quiet periods don't get
nuked.