[FIXED JENKINS-52114] Make success/failure use parallel stage result
This changes `success` and `failure` `post` conditions for stages to
look at whether there's an explicit error passed to them, rather than
just looking at build status. This doesn't apply for top-level `post`,
and has some special logic to not fire `success` if the build status
is anything but `FAILURE`, and to not fire `failure` if the error is a
result of aborting the build.

There is one potentially problematic behavior change as a result - if
you set `currentBuild.result` to `FAILURE` explicitly within a stage,
`success` will still fire, and `failure` will not fire. There's no way
to tell that this particular stage explicitly set
`currentBuild.result` vs another stage running in parallel having an
actual error.