[FIXED JENKINS-53804] Call post conditions as early as possible
Specifically, try to catch errors from directives and execute the post
conditions for the build/stage right afterwards, while we're still
inside any enclosing directives - i.e., still in top-level `agent`
when a top-level `credentials` explodes.

Also reworked post handling in general for stages when I realized that
an error in a directive before execution of a stage's steps or nested
stages would actually result in `post` for the stage not being run at
all, which is not exactly ideal.