[FIXED JENKINS-53734] Allow parallel within top-level sequential
We'll still reject any attempt to use `parallel` within another
`parallel`, but something like the below would now be allowed.

```
stages {
  stage("foo") {
    stages {
      stage("bar") {
        parallel {
          stage("a") {
            ...
          }
          stage("b") {
            ...
          }
        }
      }
    }
  }
}
```

Note that we're not expecting any visualization change for this - Blue
Ocean does not have the ability to visualize nested parallels, nor do
we expect it to any time soon, so this'll just show up there as a
sequential stage.