[FIXED JENKINS-54919] Process agent configuration on the fly
This lets us reference environment variables (well, sorta - you can
reference an environment variable defined at the parent level in a
stage agent, but not an environment variable defined at the same
level) and more immediately, this will allow referencing parameters
that have just been added to the Pipeline in this build (well, as long
as they have default values set).

Basically, this makes Agent work more like StageConditionals by
wrapping the map that defines the agent in a closure, and evaluating
that closure at the time we actually need to use it, rather than
baking the values in when we instantiate the runtime model.