Add test for label resolution from parent  env
JENKINS-43911 is still failing because the environment isn't preset when
the agent label is evaluated. However, if the stage is inside another stage
the label should be able to pick up the environment from the parent.

This is not espcially interesting but is worth having a test to cover it.