First cut at wrapped closures
This addresses the new failures by carefully pulling specific closures back into the WorkflowScript space.

With this change the 300 cell matrix is failing again, likely due to the number of statements being pulled into initialization.
However now that we have variables to hold duplicate items in we might be able ot be this by reusing steps delegates instead of creating a new one for each stage.