Allow ModelParser to be run outside of a Jenkins instance
I'm working on tooling for converting Jenkinsfiles to JSON outside of
a Jenkins instance, and there are a number of things that break as
things stand now. This fixes those, by allowing for providing a
`ModelValidator` implementation and a `DescriptorLookupCache` instance
to `ModelParser`, and uses a fallback for the zero-arg agent model map
if run outside of Jenkins.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>