fix(pct): Add the optional matrix-project dependency in test scope
As per https://github.com/jenkinsci/bom/blob/master/CONTRIBUTING.md#failed-to-load---update-required

This message usually shows that an optional dependency of a plugin is not being updated by the Plugin Compatibility Tester (PCT) and that update is needed by another plugin. This is a known issue in PCT. Workarounds include:

Add the optional dependency in test scope to the affected plugin. The workaround leaves an unnecessary test dependency in the affected plugin in order to avoid the issue
Pin an older version of the optional plugin on older LTS lines if the issue is not visible in the weekly line. This only works if the specific issue is not visible on the weekly line.
Some pull requests that add the optional dependency in test scope include:

Customer folder icon add test dependency on config-file-provider and issue 280
Mark trilead API as test even if not needed, followed by remove trilead
A pull request that pins an older version of the optional plugin is:

Pin htmlpublisher 1.36 for older lines